requirements.txt
mcp
langchain-mcp-adapters
# langchain-mistralai
langchain-openai
langgraph
python-dotenv
streamlit
fastapi
uvicorn[standard]
boto3
botocore